Add webpack config
